remote-control-gtk: Added new input-source-frame to audio panel.
The api has been changed to support input sources. Therefor we have added a
new frame to the audio panel.
To say consistant with the names we also renamed various items and there
handlers. This results in a cleaned up of the init code where we create the
different frames on the panel. So we are now more generic than before.