server: voip: Support SIP domains.
This is currently a hack. The RPC API does not explicitly allow a domain
to be set, so this commit adds a check to see if a domain is encoded
within the username parameter. If none is given, the behavior remains
unchanged and the host parameter is used as implicit domain.

Note that this is really only a hack. To solve this properly, the
medcom_voip_login() function should be enhanced to include a domain
parameter. This can probably be fixed together with the pending API
change that passes a medcom_voip_account structure to the login
function.