gtk: Rename medcom-client -> remote-control.
The project should actually be independent of the platform used. There
are still places where the medcom prefix is still in use, most notably
in the protocol definition, but those should all go away eventually.

This commit also fixes the loading of the glade user-interface
description file, which has never worked before for the installed
binary, because only the current directory was being searched.