remote-control: Add support for a configuration file.
RDP connection information is now loaded from the remote-control.conf
configuration file. If no hostname key is defined for the rdp group, no
attempt will be made to connect. If any of the username or password keys
are undefined, it will default to MTXXXXX, where XXXXX is a 5 digit
decimal representing the device's serial number (usually the last 2
octets of the MAC address in decimal).

Note that this scheme should eventually be made more generic to allow
other platforms to hook into it and replace the generated username and
password with something platform-specific. The MT prefix was chosen
specifically for the Med@tom platform.