remote-control: Conditionalize D-Bus support.
D-Bus support is not implemented yet, so there's no need to request bus
name ownership. Eventually this should either be implemented properly or
removed completely.

Note that there may actually not be a need to use D-Bus at all. At the
time this initial support was written it was contemplated to use D-Bus'
bus activation mechanism to automatically start remote-control.
Something similar may still need to be done at some point, but D-Bus may
not be required.