server: Initialize ALSA poll timeout earlier.
In the (unlikely) case that the thread is run before the timeout is
initialized, the call to poll(3) may get a timeout of 0. This may not be
a real problem, since poll(3) will just return immediately in that case,
but it is still better to be consistent here and have the timeout
initialized before the thread is even created.