remote-control: Made video playback backend configurable.
On tegra2 based systems we have no HW accel. video decoding support in VLC.
But we have support for Gstreamer and OpenMAX.
This patch allows to specify the backend that shall be used to show video on
the device. For now we have only support for VLC and stubs for GStreamer and
Null backend.