server: Implement event manager as GSource.
This allows the event manager code to run within the main thread and
works around any multithreading issues. Since the event manager source
becomes a child source of the remote-control source, it will
automatically be destroyed when its parent is.

Note that the event manager seems to work fine in a separate thread, but
other components don't, so it makes sense to keep multithreading to a
minimum.