server: Only transmit the actual LLDP frame.
Even if a large buffer is specified in the request, only the actual LLDP
frame should be transmitted within the response. On error, or if no data
is available, nothing is transferred at all.

Note that this is really only a workaround for a problem that may lie
deeper within librpc. Perhaps the recv() call does not return the
complete payload because the TCP packets are fragmented if they exceed
the MTU. This is speculation, however, and needs further investigation.