Export the client remote-control library.
This commit makes the client remote-control library's exported interface
independent of the stubs generated from the protocol definition. Note
that this needs some more work, because the current interface is not an
abstraction, but rather duplicates type definitions from the generated
stub in order to get rid of the dependency. Eventually, the library
implementation should make sure that the types are correctly mapped.