remote-control-client: Use lookup tables for parsing.
This commit uses lookup tables to do the parsing of mixer control names
and mixer input sources. The code isn't really much shorter, but it may
be easier to extend in the future (though this is not likely to happen
very often, perhaps not even at all). Sören wanted this conversion, so
even if it's not useful, then it may at least be a nice present.