remote-control: Hide the mouse cursor.
This commit adds code to always hide the mouse cursor. Since there is
normally no mouse connected, the mouse cursor is not needed and only
clutters the screen. Eventually this code should be modified to display
the cursor if a mouse is connected for debugging purposes.

Note that this commit requires a patched xfreerdp with the --kiosk
command-line option.