server: Fix initial interrupt processing.
Only changes to the interrupt status are propagated via an RPC event
notification. The RPC client needs to explicitly clear an interrupt
before another change can occur. Since the interrupt status is
already initialized before a connection is established, the RPC event
cannot be transmitted, and interrupts won't ever be cleared, which
prevents further interrupt status changes from occurring.

In order to fix this, the current interrupt status always needs to be
transmitted once when a connection is established, so that the client
can process and clear the interrupts.