remote-control-client: Rewrite command handling.
This commit is a pretty much complete rewrite of the command-handling
code. The new implementation no longer uses libsh, which is very
restrictive in some regards. For example, it is not possible to pass an
arbitrary number of arguments in a command with libsh. A subsequent
commit will introduce a new RPC call which allows executing an arbitrary
command and therefore needs to be able to take a variable number of
arguments.