remote-control: Prevent website from "breaking out".
Some websites do some ugly stuff with Javascript, like this (taken from
http://www.adac.de):

	if (window!=top) {
		top.location=window.location;
	}

to break out of frames. In order to allow the browser to be based on
HTML and Javascript, needing iframes to embed other websites, this
commit prevents the URL of the topmost frame from being set to a
different base URL than the current one.

This effectively restricts the URL of the topmost frame to be set to
anything but subpages of the first loaded URL.