Update package check for gstreamer.
In order to build gstreamer backend corretly we need a additional package,
because the required header file is located in gstreamer-plugins-base. But the
in order to link we also need gstreamer-interfaces.
Note that gstreamer-plugins-base provides the required files but only the .pc
file from interfaces provides the correct library path.