Port to librpc 0.4.6 for interface prefix support.
This new version of librpc supports the concept of interface prefix,
which is just a string that is prepended to all generated function and
type names. An uppercase version is prepended to enumeration constants.

In addition, several macros (RPC_STUB, RPC_IMPL, RPC_TYPE and RPC_MACRO)
can be used to reference prefixed functions, types and constants using
only the base name. This makes the resulting code prefix independent, so
that a prefix change requires no additional modifications.

Furthermore this change allows getting rid of all instances of the
medcom_ prefix, because that prefix is no longer valid (remote-control
runs on other platforms as well).