server: Make SmartCard/RFID backend selectable.
The configure script now takes a --with-smartcard-backend parameter that
can be used to select the desired backend for SmartCard and RFID access.
Note that currently only the dummy and libsmartcard backends are
supported, which don't provide any way to access contactless smart
cards. The dummy implementation will therefore be selected as the RFID
part of the backend.