server: Fix a segfault in the libccid-based smartcard backend.
The mutex used to serialize access to the CCID device needs to be
initialized before the thread is started, otherwise the thread may
actually start using the uninitialized mutex.

This was a typical heisenbug. It only occurred very rarely and
disappeared when run under a debugger. It was finally successfully
tracked down by setting ulimit to unlimited, causing the shell to dump a
core file when segfaulting. The resulting core dump was then run in the
debugger to find the exact location of the fault. From there it was
rather easy to diagnose and fix.