server: backlight-medatom: Use DPMS to enable/disable.
The graphics hardware on the processor boards has a pin that can be
controlled via DPMS. The backlight controller in the FPGA uses this pin
to enable or disable the backlight. Disabling and enabling also clears
the internal error state, which is why using the DPMS pin is preferred
over setting the brightness to 0 or 255 respectively. Otherwise the
backlight may go into an error state and never get out again.