Add handset subsystem.
The subsystem currently provides a dummy backend and a backend for the
PT-IP. Note that the PT-IP backend will fail miserably when no PT-IP is
connected. One problem will be that it tries to connect to the PT-IP,
using a TCP connect(), which will take rather long to timeout. After the
timeout, though, remote-control will refuse to start because the
connection cannot be made. So the only way for this to work on devices
with no PT-IP connected is to use the dummy backend.

Obviously this should be fixed so that the same root filesystem image
can be used whether or not a PT-IP is connected. One possible fix would
be to use select() with the writefds parameter containing the connecting
socket and the timeout parameter set to the given timeout. select() will
wait until the connect() can succeed. If a connection cannot be
established, the code should fallback to just not handling any PT-IP
commands, or return ENOTSUP, ENODEV or something similar.