server: Boost linphone playback volume.
The speaker of pt(no)ip is quite low. So we can't hear the remote caller realy
badly and since we have all the relevant channels from mixer already at max
the only chance to get more volume is to boost it in software.
Linphone already has a filter plugin to control the volume which we can use.
So this patch will set the internal boost to max.