server: Reworked gstreamer implementation.
Since the playbin based pipeline does not work properly we needed to implement
a manual pipeline. The current implementation does not work as expected and
did not fit our needs.
So we reimplemented a new manual pipeline for tegra based systems. Together
with changes on the omx plugins this requires a lot of changes on code
required for the user interaction.

Another big point is that this code also needs to run on 16:9 medcom and not
only plutux. Therefor we need to detect the used display used for the output.
So we use xrandr to detect the connected displays. If HDMI is connected we
need to change the pipeline in order to drive the audio to the correct output.