voip: Fix structure packing on !32-bit.
The voip_login_options structure contains pointers, which on 32-bit
architectures are 4 bytes, while on 64-bit architectures they are 8
bytes. Obviously this needs to be accounted for if transmitted via
network.

This really wasn't my fault. The structure was supposed to tunnel the
account information using the regular RPC mechanism. The initial
implementation had a bug which was fixed completely the wrong way. Now
we'll have to live with it forever.