server: Add queue for handset events.
This commit fixes a bug where handset events would be lost because of
missing buffering. If the handset sent a subsequent event before the
previous event was processed by the client, the old event would simply
be overwritten. The new implementation uses a queue (of currently
unlimited length!) to buffer events and reraises interrupts as long as
there are pending events.