server: Fix window visibility management.
For RTSP streams, the MediaPlayerPlaying event is emitted before the
stream is parsed. Therefore the callback will not have enough
information to tell whether the window should be displayed or not
(i.e. whether there are any video streams or not).

This can be fixed by using the new MediaPlayerVout event, which is
emitted when the number of video outputs required by a stream changes.
Currently this needs a patch against VLC, but it should be included in
the upcoming 1.2.0 release.