server: Keep a reference to the event-manager.
In order to keep the event-manager around until the very end, a
reference is explicitly kept in the remote-control context. Otherwise,
the destruction of the remote-control GSource causes the early
destruction of the child event-manager GSource, which in turn causes a
crash on shutdown.

This bug was reproduced by terminating the remote-control daemon while
a VoIP call is in progress. With this patch applied, the bug is gone.