server: net: Implement primary channel.
This commit introduces the concept of a network channel to share common
code among the primary and secondary channels. The primary channel is
typically used to communicate with an authentication server (like
Medial), while the secondary channel is used to communicate with the
wall-plug.

The primary channel can optionally be configured at runtime, while the
secondary channel is hard-coded.