server: modem: Add support for analog modems.
This commit takes advantage of the new libmodem features that support
modems which are connected directly connected to the microphone and
speaker.

In order to support both configurations with a USB modem or an analog
modem, the USB modem is probed first and if it does not exist, the
default serial port for the analog modem is tried.