server: Fix and extend primary interface lookup.
It can happen that a call to nl_connect() returns with an NLE_EXIST
error, which is basically the libnl equivalent of EADDRINUSE returned by
the call to bind(). While it is unclear why this happens, a workaround
is to retry with a newly generated local port until the call no longer
fails with NLE_EXIST.

Furthermore, if no default route is found (because perhaps the DHCP
configuration is not complete yet, or the device is not connected to a
network), fallback to using the first ethernet-type interface (i.e. that
has an ARP type of ARPHRD_ETHER).