server: Add radio support for gst media player.
This is another ugly hack. All uris containing "-radio" will be assumed to be
audio only streams, for all other we assume we have audio and video.
Since we had never radio before we need to hide the controlling window in this
case and show it when we have video.

This hack duplicats a lot of code. We need to switch back to autopipe as soon
as possible.