remote-control: Use GLib Unix signal sources.
Since we currently only handle SIGHUP, SIGINT and SIGTERM signals, we
can use GLib's Unix signal sources to take appropriate action. This has
several advantages: 1) it gets rid of a race condition where the signal
wouldn't stop the program when received before the main loop was started
and 2) it is safer because the actual handler is not run in the
asynchronous context of a signal handler.