server: libmodem: Add proper locking.
This commit adds proper locking for modem operations. Since polling of
the serial line is done in a separate thread, processing in that thread
needs to be locked against operations in the main thread.

Note that eventually this should all be moved into the main thread by
providing a GSource that can be integrated into the GLib main loop.