server: voip: linphone: Emit disconnect event on error.
If a call error is detected during outgoing calls, emit a disconnect
event to notify the user-interface. This fixes a bug where the user-
interface would not react if a called number was busy. Since linphone
doesn't have any built-in handling for this case either nothing at all
would happen, leaving the user guessing what could be wrong.