server: Cleanup sysfs-based backlight backend.
This commit corrects the common misconception that the value written to
the bl_power attribute of a backlight-class device should be 0 to turn
power to the backlight off and 1 to turn power on. However, 0 is
actually the numeric value for FB_BLANK_UNBLANK, which is the power
state used to turn power on, while any other value usually turns the
power off. The new code uses the symbolic constants FB_BLANK_UNBLANK and
FB_BLANK_POWERDOWN instead to turn power on and off respectively.

Furthermore, this commit cleans up some issues and inconsistencies in
the code.