Allow building against Gtk+ 2 and 3.
This can be configured by the --with-gtk configure switch, which takes
either a 2.0 or 3.0 argument. Note that when building against either
version, all other components used by remote-control or the standalone
browser (remote-control-browser) need to be built against the same Gtk+
version. The only component is currently gtkosk.

Note that we should eventually switch to Gtk+ 3 completely and get rid
of the #ifdef'ed code.