Do not treat warnings as errors by default.
Sometimes dependency updates (such as GLib) cause the build to produce
warnings (because of deprecations). When compiled with -Werror this
causes the build to fail needlessly.

While it is true that the software should eventually be updated to not
use the deprecated functions, it may be inconvenient to have to fix it
right away.