server: libmodem: Do not hang up automatically.
When the modem sends the busy detection event (<DLE> + b), the handling
is currently done in a separate thread. However, we cannot terminate the
call at that point because changing the state must always be done from
the main thread to avoid deadlocks. Instead, an event is emitted to
signal the end of the call and it is left to the user-interface to take
appropriate action.

It also seems that modems sometimes loose their temper and start to emit
DLE-escaped data (sometimes even random, non-escaped data it seems). So
ignoring these may be even better. Unfortunately there doesn't seem to
be a way to completely disable event reporting. An alternative, while
time consuming, would be to beef up the parser to cope with random data.