server: modem: Integrate with GLib main loop.
This commit converts the modem backends to use a GSource to integrate
with the GLib main loop. Among other things, this allows us to get rid
of the extra thread that was used to poll data from the modem serial
port. Needless to say that this also makes the code a lot simpler.

Since the modem structure now wraps a GSource, all memory management is
done via the GSource and the corresponding callbacks. Hence, the old
modem_manager_free() function is no longer needed.

This commit requires functionality only available in libmodem 0.2.2 and
later.