server: ucm: Minor cleanups.
Be more explicit, since the behavior of strcmp is a little bit odd. If the
strings are the same, it'll return zero, which can also be thought of as
"false".  Not-false is "true", so (!strcmp()) will be true if the strings
are the same.

It is not false to write (!strcmp(foo, bar)) but it is always a little bit
confusing to say NOT but meaning YES.