server: Implement GPIO API.
This commit implements the GPIO API (formerly known as Just-A-Board) on
the server side. Note that this requires rather invasive changes to the
event-manager and remote-control objects.

Also this API is not very well designed and is not at all flexible. For
instance, there is no possibility to control the direction of a given
GPIO and the number of available pins is completely unknown. If this is
ever redesigned, those issues should be addressed and the API should
generally mirror the functionality of a generic GPIO controller.