remote-control: Add JavaScript input device callback.
This callback can be used by JavaScript code to register a function to
be called when events are detected on Linux kernel evdev devices.
Currently this is hard-coded to only monitor the sx8634 capacitive touch
controllers found on Avionic Design Medcom-Wide platforms.

Note that this commit contains some cleanups of the older JavaScript API
that controls the mouse cursor and reads events from the IR receiver.
Eventually, those APIs should be refactored into proper JavaScript
objects.