server: gstreamer: Increase plugin priority.
In commit 86760d4 we increased the priority for the ffmpeg mp3 decoder plugin
from NONE to MARGINAL. This was not enougth to work with all streams properly.
So we need to increase the priority again by one click.

Webradio still works with this setting. This has been verified with the
following uri's:
	mms://dms-cl-022.skypro-media.net/basilisk-128
	http://stream.hoerradar.de/psr128