remote-control: Allow modem backend configuration.
This commit passes the global configuration to the modem backend so that
it can use the configuration to select a modem device, VLS mode and ATL
volume. An example configuration snippet is given below:

	[modem]
	device = /dev/ttyS0
	vls = 13
	atl = 3

This can be stored f.e. in the /etc/remote-control.conf.d/modem.conf
configuration file.