remote-control-browser: Make language configurable.
This commit allows the languages accepted by the browser (as specified
in RFC2616, 14.4 Accept-Language) to be configured. The setting is taken
from two sources: the new browser.conf configuration file (typically
found in the /etc directory) and the command-line (--accept-language
option). If both are specified, the command-line option overrides the
configuration file.