bin/remote-control: Add check if private data is valid.
It can happend that the objects context has been changed. In this case we
would end up in the middle of nowhere if we not check.
One possible way to create such a scenario is, to to assign the function in
JavaScript to another variable and call it, e.g.

	var foo = AvionicDesign.Cursor.moveTo;
	foo(10, 10);