remote-control: Add D-Bus Watchdog support
This commit allows the remote-control process to be monitored by the
D-Bus watchdog daemon. Note that even when support is enabled with the
--enable-watchdog configure option, the watchdog will only be activated
if a corresponding entry is found in the configuration file.

Example:

	[watchdog]
	timeout = 13000

This will register the remote-control process with the D-Bus watchdog
with a timeout of 13 seconds and send a ping to the watchdog three times
during within the interval (every 4 seconds).