remote-control: Pass GLib main loop to windows
Instead of passing the GMainContext, which is enough for windows to add
GSource objects to the main loop, the main GMainLoop is now passed. A
GMainContext object can always be obtained from its GMainLoop parent.

A follow up patch relies on this behaviour in order to terminate the
main loop.