remote-control: Terminate when RDP session is closed
When an RDP session is closed, remote-control should be restarted to
avoid any potential synchronization problems with the remote session.
This fixes a bug where the application running in the remote session
starts a browser locally, which can no longer be closed when the remote
session is terminated and reconnected.