server: modem: libmodem: Restore log message and error code
Log messages should generally not be changed unless there is a very good
reason. This is important because somebody might be using the specific
format of the log message to search logs for errors. Changing the
message for no reason may cause such automatic tools to fail
unnecessarily.

Also the error code was mistakenly changed in a previous commit, which
would result in broken error handling in the caller.