Add configure option for preferred audio languages.
In remote-control.conf one can set a string list for the preferred
audio-languages. Section is media-player, variable name is preferred-languages
and seperator is ";", the priority is highest on the first list entry and
reduces continuosly. This variable is used in gstreamer backend only so far.
The backend automatically selects the audio track which matches best.