gstreamer: Go to PAUSED instead of READY while resolution changes.
When a fixed fullscreen resolution is used the pipeline is PAUSED while the
xrandr call is emitted. This prevents the video output plugins from running
into errors. Previously we used READY state, as PAUSED was not working
properly with playbin. As this is fixed with playbin2 we use PAUSED now.