Strip armv7l out of user-agent.
On tegra-based devices webkit will include armv7l as processor type in the
user-agent by default. While this is generally ok, it is abused by several
websites to detect mobile devices and show mobile versions of webpages. As we
don't want to show the mobile versions, we strip armv7l out now.