media-player-gtk-gst: Retry changing resolution on failure.
When XRRSetScreenConfigAndRate returns a RRSetConfigInvalidConfigTime failure,
the selected values for resolution and/or rate where not correct. In this case
we need to re-query the currently possible settings from xserver and update
our selection.

This will fix a issue where we try to change the resolution but the xserver
does not except the change and we end up with a state where either the
screen resolution does not match the video output window. Or the screen does
not return to the expection resolution.