media-player-gtk-gst: sync state change optional
Waiting synchronously for a state change to complete is not a good idea in the
most cases, as it will block the application for the time waiting. It should
only be used, when it is really necessary to ensure that a state change
completed before further action can be done.