bin/remote-control: Add javascript support for RS232 LCD-TV control.
This will extend remote-control's javascript api with a new class LCD.
It allows us to control LG LCD-TV's using there rs232 port.

This implementation is very basic and may change in the future. As
requested by the bosses we also need to allow to send raw data via the
rs232 port.