media-player-gtk-gst: Use stored resolution for restore.
If we change resolution for fullscreen, we can not use the current resolution
to check if we what to restore the original resolution. Instead we need to
store the resolution at startup and compare the new resolution with the stored
one.

This solves a problem which can occure, if the resolution for fullscreen is
smaller then the resolution we use within the original screen resolution.
E.g. native resolution is 1920x1080 we use 1280x720 in fullscreen but 1400x768
for preview.