gpio-sysfs: load settings from remote-control.conf
This adds a gpio group to the remote-control.conf which has to be used to set
the gpio addresses used by the sysfs backend from now on. As the previous
settings were hardcoded for medcom-wide (Tegra based) the new configuration
options have to be added to the platform configuration before this version can
be used.