media-player-gtk-gst: Add MRL proxy option support.
This is a fix for FS#1604. It will add support for HTTP stream where the
URL contains a MRL like optional proxy
(e.g. http://www.somewhere.com/stream.mp3 :http-proxy=proxy-server).

So we need to parse the url first, as we do it already for audio-language,
and pass the proxy as soon as the http-source is ready. It will also allow
us to use the same stream URLs for VLC and gstreamer backend.