common: Move log utilities to central location
The log utilities can be useful not only in the remote control server
implementation but also in other components, so they are moved to a more
central location. While at it, a new pr_debug() macro is added, which
takes into account the pr_fmt() macro to simplify adding a prefix to log
messages.

Signed-off-by: Thierry Reding <thierry.reding@avionic-design.de>
Reviewed-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>