common: Add GLib API for sysfs and GPIO support
Sysfs operations are already quite common in remote-control and more
will likely be required in the future. Factor out some of the
functionality and add a GLib-style GPIO API on top.

The GPIO API will be used subsequently to control the power GPIO for I2C
SmartCards and could be extended to replace the sysfs accessors provided
by libgpio for the sysfs GPIO backend.

Signed-off-by: Thierry Reding <thierry.reding@avionic-design.de>
Reviewed-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>