server: smartcard-i2c: Implement power GPIO support
The power to the SmartCard slot can be controlled by a specified GPIO.
This commit adds support for obtaining information about the GPIO and
its configuration from the configuration file. Furthermore, the I2C bus
and slave address by which the SmartCard is accessed is also made
configurable.

Finally, the remote-control.conf documentation is updated to describe
the new configuration options.

Signed-off-by: Thierry Reding <thierry.reding@avionic-design.de>
Reviewed-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>