voip: linphone: Treat 0 as invalid expiration time
If the registration-expiry key isn't found in the configuration, it will
default to 0. Instead of adjusting this to the lower bound (30 seconds)
of the allowed range it should be handled as specifying an invalid value
and therefore the linphone default should be used.

Note that the GKeyFile API will return 0 if the registration-expiry key
isn't found in the configuration, so the extra check for the key's
existence could just as well be removed.

Signed-off-by: Thierry Reding <thierry.reding@avionic-design.de>
Reviewed-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>