backlight: medatom: Add sysfs support
The Medatom-Wide has an improved BIOS which properly initializes the
graphics driver's backlight controller. This allows the brightness of
the display to be controlled via the sysfs backlight interface. If the
backlight controller is not exported via sysfs, it is assumed that the
legacy I2C-based controller will be available like on classic Medatom.

Signed-off-by: Thierry Reding <thierry.reding@avionic-design.de>
Reviewed-by: Julian Scheel <julian@jusst.de>