voip-linphone: Remove all registrars on logout.
This can help to workaround a linphone issue. When voip-logout is called
and the registration was not successful. Linphone does _not_ stop to send
out register requests.

When the controling application starts to call login, logout, login,
logout, ... and so on. Linphone sends register requests for each previously
called login.

Signed-off-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>
Reviewed-by: Thierry Reding <thierry.reding@avionic-design.de>