remote-control: Run RPC commands in separate thread
Recent versions of WebKitGtk or any of its dependencies show different
behaviour in how synchronous XMLHttpRequests are handled. Older versions
used to work properly when they were run in the same main loop as the
RPC calls but recent versions don't. This can easily be seen by running
the certification software, which renders in the WebKit process that
runs in the default main loop and uses XMLHttpRequest objects to connect
to rum, which in turns converts the requests to RPC calls handled in the
same main loop.

On recent versions of WebKitGtk this deadlocks, because the completion
of the XMLHttpRequest relies on the RPC call finishing, which won't
happen since the XMLHttpRequest blocks the main loop. The ugly solution
implemented by this commit is to run the RPC call handling code in a
separate thread and main loop to allow both the XMLHttpRequest and the
RPC call to run in parallel.

Signed-off-by: Thierry Reding <thierry.reding@avionic-design.de>
Reviewed-by: Julian Scheel <julian@jusst.de>