remote-control-browser: Fix WebKitWebView access from notebook
When using Gtk+ 2.0, the WebKitWebView widget is wrapped in a Katze-
Scrolled widget to enable kinetic scrolling. This is no longer needed
with Gtk+ 3.0 which has builtin kinetic scrolling support.

Unfortunately WebKitGtk does not use the scrolling implementation
provided by Gtk+ but rather reimplements it. This, however, prevents the
Gtk+ kinetic scrolling from being used and furthermore does not allow
embedding the WebKitWebView widget within a Gtk+ scrollable widget.
Effectively this means that there is currently no way to enable kinetic
scrolling on WebKitGtk with Gtk+ 3.0.

This commit fixes up the way the WebKitWebView widget is retrieved from
the notebook. If Gtk+ 2.0 is used, the widget needs to be obtained from
the parent KatzeScrolled widget. For Gtk+ 3.0 a simple cast is enough.

Signed-off-by: Thierry Reding <thierry.reding@avionic-design.de>
Reviewed-by: Julian Scheel <julian@jusst.de>