remote-control-browser: Fix crash on termination
When the browser is closed while a page is still loading, the WebKit
widget will try to send progress updates to the URI entry widget which
may already have been destroyed, therefore causing a crash. This patch
solves this by stopping any loads that are still in progress before
destroying the top-level window.

Signed-off-by: Thierry Reding <thierry.reding@avionic-design.de>
Reviewed-by: Julian Scheel <julian@jusst.de>