remote-control-browser: Refactor querying of the load status
The load status is queried at various points in the program, but it is
only used to determine whether the page is currently loading or not. A
new function webkit_web_view_is_loading() is introduced to replace any
open-coded equivalents.

As a side-effect the loading flag that was kept in the WebKitBrowser's
private storage can be removed. The usage of this flag was broken in
that it didn't keep the loading status on a per-widget basis. The new
function always ensures that the status of a specific widget is used.

Signed-off-by: Thierry Reding <thierry.reding@avionic-design.de>
Reviewed-by: Julian Scheel <julian@jusst.de>