remote-control-browser: Fix initial keyboard focus
This commit makes the current WebKit widget grab the keyboard focus.
Otherwise the toolbar would get the keyboard focus by default, which
would cause the back button to be highlighted, which doesn't make a
lot of sense in a touchscreen-oriented application.

Besides this visual fix it actually prevents a bug that can be triggered
by using the gtkosk onscreen keyboard. What happens is that if the back
button has the keyboard focus, pressing space on the OSK will activate
the button, which in turn will cause the OSK to loose the focus and not
receive the release event. A separate fix for gtkosk has already been
implemented.

Signed-off-by: Thierry Reding <thierry.reding@avionic-design.de>
Reviewed-by: Julian Scheel <julian@jusst.de>