remote-control-browser: Make maximum number of pages configurable
The default maximum number of pages that can be opened is 8. This commit
allows the default to be overridden via an option in the browser.conf
configuration file.

While at it, the logic to enable or disable the tab add and delete
buttons is simplified and actually fixed. Previously if only a single
page was allowed to be opened, the delete button would remain active.

Signed-off-by: Thierry Reding <thierry.reding@avionic-design.de>
Reviewed-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>