browser: Fix crash due passing uncheck format string.
When we pass a url to the watchdog_set_message, is is possible that
g_strdup_vprintf will cause a unchecked access, which will lead to
program abort.
Because we do not use the format option right now, we can drop this
feature. Also we limit the length of the message to 64 characters.
The message is only used for debugging and has no further relevance.

Change-Id: I0e1509cb6f8f3101ba4a0b5a9a360b4e3394b312
Signed-off-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/112
Reviewed-by: Dirk Leber <dirk.leber@avionic-design.de>