browser: Fix watchdog issue.
We pass the current URL to the watchdog, so we have a indication on which URL
the watchdog was triggered. Because the URL can include unicode characters
encoded starting with %. This can lead to problems when the URL string is
used afterwards in functions such as g_strdup_vprintf.
Because we don't need the URL we can replace the % with a # to avoid issues.

Change-Id: If90e9f82a7b16143d0bdd114a878a6653d22f1cb
Signed-off-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/118
Reviewed-by: Dirk Leber <dirk.leber@avionic-design.de>