server: audio: Add exception to control usb-handset volume.
The usb-handset is actually an additional soundcard. So we need to control a
different mixer-control on this card, when we are in a audio-state where the
handset is used.
This adds a kind of a hack, but we do not have a better solution, and as it is
already in production, we need to commit this.

Change-Id: I912019d6d66da80ccc6ef39ac3864a752c85c8c7
Signed-off-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>
Tested-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/97
Reviewed-by: Dirk Leber <dirk.leber@avionic-design.de>