server: audio: Workaround mute when setting volume.
For a unknown reason it can and does happen (not always, only sometimes) that
the control were we set the volume on, gets muted. It looks like the mute is
set right before setting the volume.

The issue appears since kernel update from l4t-r15 to l4t-r16.2. So it is most
likely a kernel bug, and should be fixed there. But since this workaround is
already spread out into the field, we need to commit this.

Change-Id: I4853e01c66ca14cb8a69b26aa1ebb3f555a17053
Signed-off-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>
Tested-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/98
Reviewed-by: Dirk Leber <dirk.leber@avionic-design.de>