server: voip-linphone: Logout on shutdown.
When destroying and cleaning up the linphone object, we now also deregistering
from the server. This should solve also problems with tcp connections.

When using tcp for the server connection, the socket on server side is not
closed in time, when we do a restart of remote-control. So that a not properly
configured sip server does criticise that the port is still in use. The
deregistering on shutdown, this problem should be solved.

Change-Id: I1f629addc30983225a06480e9b41ec943d3da798
Signed-off-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/120
Reviewed-by: Dirk Leber <dirk.leber@avionic-design.de>