server: lldp: Make lldp interface configurable.
When we started to use a bridged interfaces we can no longer use the
interface with the default route, because it can be the bridge. But lldp
does not work on bridges.
To deal with this, we now allow to specify the interfaces to use for lldp
monitoring via configuration file.

Change-Id: Ib57e35c174f4e76178f77de4c8caf2279d522283
Signed-off-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/121
Reviewed-by: Dirk Leber <dirk.leber@avionic-design.de>